cleanup and move aw widget to own folder
