+++ src/iced_widgets/mod.rs
@@ -1,9 +1,7 @@
 
 
 
-pub mod card;
 pub mod checkbox;
-pub mod menu;
 pub mod mousearea;
 pub mod mousearea_table;
 pub mod scrollable;
