+++ src/ipg_widgets/mod.rs
@@ -5,7 +5,6 @@ pub mod helpers;
 pub mod ipg_button;
 pub mod ipg_card;
 pub mod ipg_checkbox;
-// pub mod ipg_color_picker;
 pub mod ipg_column;
 pub mod ipg_container;
 pub mod ipg_date_picker;
@@ -15,7 +14,6 @@ pub mod ipg_image;
 pub mod ipg_menu;
 pub mod ipg_modal;
 pub mod ipg_mousearea;
-// pub mod ipg_pane_grid;
 pub mod ipg_pick_list;
 pub mod ipg_progress_bar;
 pub mod ipg_radio;
@@ -28,7 +26,6 @@ pub mod ipg_space;
 pub mod ipg_svg;
 pub mod ipg_table;
 pub mod ipg_text;
-// pub mod ipg_text_editor;
 pub mod ipg_text_input;
 pub mod ipg_theme_colors;
 pub mod ipg_timer;
